requirements.txt
"""
Python 3.9+
"""

# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0

# LLM & AI
langchain==0.1.0
langchain-openai==0.0.5
openai==1.3.0

# Embedding Model (BGE-M3 - open-source multilingual)
sentence-transformers==2.2.2

# Data & Search
numpy==1.24.0
pandas==2.1.0

# Vector Database (choose one)
chromadb==0.4.0
# pinecone-client==3.0.0  # Optional: Pinecone instead of Chroma

# Document Processing
pypandoc==1.11  # For Markdown to PDF conversion
reportlab==4.0.0  # Alternative PDF generation

# Visualization
matplotlib==3.8.0
plotly==5.18.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
requests==2.31.0
tqdm==4.66.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.0
